Move node-gui/backend to a crate
