+++ node-gui/backend/src/error.rs
