+++ node-gui/backend/src/p2p_event_handler.rs
