+++ node-gui/backend/src/wallet_events/mod.rs
